MINOR: global: Add an option to get the old listening sockets.
Add the "-x" flag, that takes a path to a unix socket as an argument. If
used, haproxy will connect to the socket, and asks to get all the
listening sockets from the old process. Any failure is fatal.
This is needed to get seamless reloads on linux.

(cherry picked from commit f73629d23a44995f84887859fcbc7d22a2227eec)
Signed-off-by: Dave Chiluk <dchiluk@indeed.com>